# Improve IncludeXmlComments performance (2)