--- a/src/Swashbuckle.AspNetCore.SwaggerGen/XmlComments/XPathNavigatorExtensions.cs
+++ b/src/Swashbuckle.AspNetCore.SwaggerGen/XmlComments/XPathNavigatorExtensions.cs
@@ -5,17 +5,29 @@ namespace Swashbuckle.AspNetCore.SwaggerGen;
 
 internal static class XPathNavigatorExtensions
 {
+    private const string EmptyNamespace = "";
+
+    internal static XPathNodeIterator SelectChildren(this XPathNavigator navigator, string name)
+    {
+        return navigator.SelectChildren(name, EmptyNamespace);
+    }
+
+    internal static string GetAttribute(this XPathNavigator navigator, string name)
+    {
+        return navigator.GetAttribute(name, EmptyNamespace);
+    }
+
     internal static XPathNavigator SelectFirstChild(this XPathNavigator navigator, string name)
     {
-        return navigator.SelectChildren(name, "")
+        return navigator.SelectChildren(name, EmptyNamespace)
                 ?.Cast<XPathNavigator>()
                 .FirstOrDefault();
     }
 
     internal static XPathNavigator SelectFirstChildWithAttribute(this XPathNavigator navigator, string childName, string attributeName, string attributeValue)
     {
-        return navigator.SelectChildren(childName, "")
+        return navigator.SelectChildren(childName, EmptyNamespace)
                 ?.Cast<XPathNavigator>()
-                .FirstOrDefault(n => n.GetAttribute(attributeName, "") == attributeValue);
+                .FirstOrDefault(n => n.GetAttribute(attributeName, EmptyNamespace) == attributeValue);
     }
 }