--- a/src/Swashbuckle.AspNetCore.SwaggerGen/XmlComments/XmlCommentsDocumentHelper.cs
+++ b/src/Swashbuckle.AspNetCore.SwaggerGen/XmlComments/XmlCommentsDocumentHelper.cs
@@ -1,0 +1,25 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using System.Xml.XPath;
+
+namespace Swashbuckle.AspNetCore.SwaggerGen
+{
+    internal static class XmlCommentsDocumentHelper
+    {
+        internal static IReadOnlyDictionary<string, XPathNavigator> GetMemberDictionary(XPathDocument xmlDoc)
+        {
+            var members = xmlDoc.CreateNavigator()
+                .SelectFirstChild("doc")
+                ?.SelectFirstChild("members")
+                ?.SelectChildren("member", "")
+                ?.OfType<XPathNavigator>();
+
+            if (members == null)
+            {
+                return new Dictionary<string, XPathNavigator>();
+            }
+
+            return members.ToDictionary(memberNode => memberNode.GetAttribute("name", ""));
+        }
+    }
+}