--- a/src/Swashbuckle.AspNetCore.SwaggerGen/XmlComments/XmlCommentsDocumentHelper.cs
+++ b/src/Swashbuckle.AspNetCore.SwaggerGen/XmlComments/XmlCommentsDocumentHelper.cs
@@ -6,7 +6,7 @@ namespace Swashbuckle.AspNetCore.SwaggerGen
 {
     internal static class XmlCommentsDocumentHelper
     {
-        internal static IReadOnlyDictionary<string, XPathNavigator> CreateMemberDictionary(XPathDocument xmlDoc)
+        internal static Dictionary<string, XPathNavigator> CreateMemberDictionary(XPathDocument xmlDoc)
         {
             var members = xmlDoc.CreateNavigator()
                 .SelectFirstChild("doc")