--- a/src/Swashbuckle.AspNetCore.SwaggerGen/XmlComments/XmlCommentsDocumentHelper.cs
+++ b/src/Swashbuckle.AspNetCore.SwaggerGen/XmlComments/XmlCommentsDocumentHelper.cs
@@ -2,24 +2,23 @@
 using System.Linq;
 using System.Xml.XPath;
 
-namespace Swashbuckle.AspNetCore.SwaggerGen
+namespace Swashbuckle.AspNetCore.SwaggerGen;
+
+internal static class XmlCommentsDocumentHelper
 {
-    internal static class XmlCommentsDocumentHelper
+    internal static Dictionary<string, XPathNavigator> CreateMemberDictionary(XPathDocument xmlDoc)
     {
-        internal static Dictionary<string, XPathNavigator> CreateMemberDictionary(XPathDocument xmlDoc)
+        var members = xmlDoc.CreateNavigator()
+            .SelectFirstChild("doc")
+            ?.SelectFirstChild("members")
+            ?.SelectChildren("member")
+            ?.OfType<XPathNavigator>();
+
+        if (members == null)
         {
-            var members = xmlDoc.CreateNavigator()
-                .SelectFirstChild("doc")
-                ?.SelectFirstChild("members")
-                ?.SelectChildren("member")
-                ?.OfType<XPathNavigator>();
-
-            if (members == null)
-            {
-                return new Dictionary<string, XPathNavigator>();
-            }
-
-            return members.ToDictionary(memberNode => memberNode.GetAttribute("name"));
+            return new Dictionary<string, XPathNavigator>();
         }
+
+        return members.ToDictionary(memberNode => memberNode.GetAttribute("name"));
     }
 }