--- a/src/Swashbuckle.AspNetCore.SwaggerGen/XmlComments/XmlCommentsExampleHelper.cs
+++ b/src/Swashbuckle.AspNetCore.SwaggerGen/XmlComments/XmlCommentsExampleHelper.cs
@@ -2,25 +2,26 @@
 using Microsoft.OpenApi.Any;
 using Microsoft.OpenApi.Models;
 
-namespace Swashbuckle.AspNetCore.SwaggerGen;
-
-internal static class XmlCommentsExampleHelper
+namespace Swashbuckle.AspNetCore.SwaggerGen
 {
-    public static IOpenApiAny Create(
-        SchemaRepository schemaRepository,
-        OpenApiSchema schema,
-        string exampleString)
+    internal static class XmlCommentsExampleHelper
     {
-        var isStringType =
-            (schema?.ResolveType(schemaRepository) == "string") &&
-            !string.Equals(exampleString, "null");
+        public static IOpenApiAny Create(
+            SchemaRepository schemaRepository,
+            OpenApiSchema schema,
+            string exampleString)
+        {
+            var isStringType =
+                (schema?.ResolveType(schemaRepository) == "string") &&
+                !string.Equals(exampleString, "null");
 
-        var exampleAsJson = isStringType
-                ? JsonSerializer.Serialize(exampleString)
-                : exampleString;
+            var exampleAsJson = isStringType
+                    ? JsonSerializer.Serialize(exampleString)
+                    : exampleString;
 
-        var example = OpenApiAnyFactory.CreateFromJson(exampleAsJson);
+            var example = OpenApiAnyFactory.CreateFromJson(exampleAsJson);
 
-        return example;
+            return example;
+        }
     }
 }