<!-- Thank you for contributing to Swashbuckle.AspNetCore!  Open source is only as strong as its contributors. -->

# Pull Request

## The issue or feature being addressed

Fixes #2164

## Details on the issue fix or feature implementation

This is a rebase and enhancement of #2443  (thanks to @stevendarby, added you as co-author).

This significantly improves the performance of the XML comment filters. The original fix #2443 created a dictionary of the members to avoid iterating the whole XPath for every single item.

I made some further improvements by replacing all usages of `Select` and `SelectSingleNode` with `SelectChildren` (added some helper extensions for this). This completely avoids XPath expressions and [the documentation](https://learn.microsoft.com/en-us/dotnet/standard/data/xml/select-xml-data-using-xpathnavigator#optimized-selection-methods) even mentions `SelectChildren` being faster.

I made the new constructors internal for now (which is also why I changed `ParameterFilter` to `AddParameterFilterInstance` etc.. DI fails otherwise). It felt wrong to me to make these public, but I can make these public if you want.

### Tests

The existing tests imho already cover this code pretty well. I only added a test case for a bug I noticed in #2443  (there was a `return` instead of a `continue` in `XmlCommentsDocumentFilter`).

### Performance

#2443 already mentions significant improvements (25s -> 4s). With another local test case I have with a 1.5MB XML file, generation time decreases from ~1900ms to ~200ms.

[#2443](https://github.com/domaindrivendev/Swashbuckle.AspNetCore/pull/2443#issuecomment-2094261461) also mentions some concern about memory consumption, so I made some manual tests with a very big (30MB) XML document file: the XPathDocument itself already loads the document completely into memory which uses about 60 MB.  Creating the dictionary creates an additional 40MB. So there is some additional memory consumption, but considering that 30MB is pretty extreme and most files should be < 1MB, the memory increase should be negligible for most applications. In my local test case with a 1.5 MB File, memory consumption after GC.Collect increased by about 1MB.