@mus65 2024-08-26
I created a benchmark in a separate branch [here](https://github.com/mus65/Swashbuckle.AspNetCore/commit/65bfd9745ca2cb757326b79cf3ef88959239af4d). As expected, the new implementation is a lot
faster, even with a small member count. The performance of the new implementation stays the same,
regardless of member count. The old implementation gets significantly slower the more members
exist. `addMemberCount` can be changed in the benchmark to try different member counts.

Old implementation with small member count (n = 79):

| Method      | Mean      | Error     | StdDev    | Median    | Gen0   | Gen1   | Allocated |
|------------ |----------:|----------:|----------:|----------:|-------:|-------:|----------:|
| Document    |  4.074 us | 0.0811 us | 0.1780 us |  4.002 us | 2.2278 | 0.5569 |  20.53 KB |
| Operation   | 12.692 us | 0.0828 us | 0.0775 us | 12.712 us | 5.0659 | 0.0610 |  47.06 KB |
| Parameter   |  6.878 us | 0.1368 us | 0.2360 us |  6.786 us | 2.6321 | 0.0076 |  24.29 KB |
| RequestBody |  4.122 us | 0.0185 us | 0.0164 us |  4.125 us | 2.2278 | 0.0610 |  20.64 KB |

Old implementation with big member count (n = 10.000):

| Method      | Mean       | Error    | StdDev   | Median     | Gen0     | Gen1   | Allocated  |
|------------ |-----------:|---------:|---------:|-----------:|---------:|-------:|-----------:|
| Document    |   541.3 us | 10.56 us | 13.73 us |   542.8 us |  78.1250 | 1.9531 |  723.66 KB |
| Operation   | 1,061.1 us | 20.70 us | 31.61 us | 1,074.7 us | 156.2500 |      - | 1453.34 KB |
| Parameter   |   489.3 us |  6.42 us |  5.01 us |   489.4 us |  79.1016 |      - |  727.42 KB |
| RequestBody |   476.8 us |  9.43 us | 16.52 us |   467.4 us |  78.1250 |      - |  723.79 KB |

New implementation with any member count:

| Method      | Mean     | Error     | StdDev    | Median   | Gen0   | Gen1   | Allocated |
|------------ |---------:|----------:|----------:|---------:|-------:|-------:|----------:|
| Document    | 2.675 us | 0.0530 us | 0.0955 us | 2.726 us | 1.8959 | 0.6294 |  17.45 KB |
| Operation   | 4.851 us | 0.0907 us | 0.0848 us | 4.869 us | 3.7842 | 0.1526 |  34.84 KB |
| Parameter   | 2.421 us | 0.0313 us | 0.0293 us | 2.411 us | 1.8959 | 0.0687 |  17.46 KB |
| RequestBody | 2.545 us | 0.0136 us | 0.0127 us | 2.545 us | 1.8158 |      - |  16.73 KB |