@martincostello 2024-08-26
Thanks @mus65 for doing those benchmarks - I'll take a closer look at them tomorrow 👍