@martincostello 2024-08-27
Thanks again for adding these benchmarks. I've pushed up a version of them in #3050 for us to use going forwards.

These are the results I get on my laptop:

### N=79

#### Base

_Pending_

#### This PR

_Pending_

### N=10,000

#### Base

```
BenchmarkDotNet v0.14.0, Windows 11 (10.0.22621.4037/22H2/2022Update/SunValley2)
12th Gen Intel Core i7-1270P, 1 CPU, 16 logical and 12 physical cores
.NET SDK 8.0.401
  [Host]     : .NET 8.0.8 (8.0.824.36612), X64 RyuJIT AVX2
  DefaultJob : .NET 8.0.8 (8.0.824.36612), X64 RyuJIT AVX2
```

| Method      | Mean       | Error    | StdDev   | Gen0     | Gen1   | Allocated  |
|------------ |-----------:|---------:|---------:|---------:|-------:|-----------:|
| Document    |   598.6 μs |  8.41 μs |  7.02 μs |  78.1250 | 1.9531 |  724.36 KB |
| Operation   | 1,105.6 μs | 19.72 μs | 17.48 μs | 156.2500 |      - | 1453.67 KB |
| Parameter   |   516.7 μs |  9.92 μs |  8.28 μs |  79.1016 |      - |  727.55 KB |
| RequestBody |   487.9 μs |  4.22 μs |  3.52 μs |  78.1250 |      - |  723.79 KB |

#### This PR

```
BenchmarkDotNet v0.14.0, Windows 11 (10.0.22621.4037/22H2/2022Update/SunValley2)
12th Gen Intel Core i7-1270P, 1 CPU, 16 logical and 12 physical cores
.NET SDK 8.0.401
  [Host]     : .NET 8.0.8 (8.0.824.36612), X64 RyuJIT AVX2
  DefaultJob : .NET 8.0.8 (8.0.824.36612), X64 RyuJIT AVX2
```

| Method      | Mean      | Error     | StdDev    | Median    | Gen0   | Gen1   | Allocated |
|------------ |----------:|----------:|----------:|----------:|-------:|-------:|----------:|
| Document    |  3.872 μs | 0.1670 μs | 0.4655 μs |  3.857 μs | 1.9150 | 0.4807 |  17.66 KB |
| Operation   | 12.341 μs | 1.2843 μs | 3.7869 μs | 14.136 μs | 3.8147 |      - |  35.17 KB |
| Parameter   |  7.158 μs | 0.3641 μs | 1.0679 μs |  7.320 μs | 1.8921 | 0.0916 |   17.6 KB |
| RequestBody |  7.953 μs | 0.3605 μs | 1.0629 μs |  8.024 μs | 1.8005 |      - |  16.73 KB |