@martincostello 2024-09-28
I have [some benchmarks](https://github.com/martincostello/aspnetcore-openapi/tree/main/perf/TodoApp.Benchmarks) I've been using to compare the benchmarks of Swashbuckle, NSwag and the new .NET 9 ASP.NET Core OpenAPI library, and there's a noticeable really improvement from the ingesting changes from this PR.

![image](https://github.com/user-attachments/assets/224a9b94-6bd8-48eb-ba16-423a99b0a5fc)

The first major improvement is from updating to .NET 9 RC1 (https://github.com/martincostello/aspnetcore-openapi/commit/7e5e365cd54751db3a9f7eacdb7d6c2c4cf114d6), then the second improvement with the big reduction in memory usage is from updating to 6.8.0 (https://github.com/martincostello/aspnetcore-openapi/commit/b4aba6567d2f114df11d3bf0fd53495b3ba393d8)

Thanks again!